revert: back to single card component and fixing SSR hydration issues
