--- opendata.swiss/ui/app/components/CardProps.ts
@@ -1,6 +1,0 @@
-export default interface CardProps {
-  title: string
-  type?: 'default' | 'highlight' | 'twitter' | 'flat' | 'universal' | 'list'
-  clickable?: boolean
-  href?: string
-}
